Refactor de las rutas del Navi
