+++ transportes-fernandez/src/data/navi_routes.js
@@ -1,9 +1,9 @@
 // Import Screens (Pages)
-import Home from '../screens/Home';
-import About from '../screens/About';
-import Contact from '../screens/Contact';
-import Services from '../screens/Services';
-import Team from '../screens/Team';
+import Home from '../screens/Navigation/Home';
+import About from '../screens/Navigation/About';
+import Contact from '../screens/Navigation/Contact';
+import Services from '../screens/Navigation/Services';
+import Team from '../screens/Navigation/Team';
 
 const navi_routes = [
     { label: 'Inicio', path: '/', component: Home },
