--- transportes-fernandez/src/screens/Contact.jsx
@@ -1,19 +1,0 @@
-import React from 'react'
-import { styled } from '@mui/material/styles'
-import Box from '@mui/material/Box'
-import Paper from '@mui/material/Paper'
-import Grid from '@mui/material/Grid'
-import ContactForm from '../components/features/ContactForm/ContactForm'
-
-function Contact() {
-  return (
-    <div id='Contact'>
-      <h1>Contact Us</h1>
-        <ContactForm />
-      <br />
-    </div>
-  )
-}
-
-export default Contact
-
