+++ transportes-fernandez/src/screens/Navigation/About.jsx
@@ -1,19 +1,16 @@
 import React from 'react';
-import MediaPicture from '../components/features/MediaPicture/MediaPicture';
-import TextBox from '../components/layout/Box/TextBox';
+import MediaPicture from '../../components/features/MediaPicture/MediaPicture';
+import TextBox from '../../components/layout/Box/TextBox';
 
-import laMancha from '../assets/original_images/regional.png';
-import ciutatVella from '../assets/original_images/nacional.jpg';
-import space from '../assets/images/space.jpg';
+import laMancha from '../../assets/original_images/regional.png';
+import ciutatVella from '../../assets/original_images/nacional.jpg';
+import space from '../../assets/images/space.jpg';
 
 function About() {
   return (
     <div id="About">
 
       <MediaPicture src={space} alt="placeholder" className='picture-container' />
-
-      
-
 
       <p>Learn more about our company and services. TONTO EL QUE LO LEA</p>
