+++ transportes-fernandez/src/screens/Navigation/Contact.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+import ContactForm from '../../components/features/ContactForm/ContactForm'
+
+function Contact() {
+  return (
+    <div id='Contact'>
+      <h1>Contact Us</h1>
+        <ContactForm />
+      <br />
+    </div>
+  )
+}
+
+export default Contact
+
