+++ transportes-fernandez/src/screens/Navigation/Services.jsx
@@ -1,13 +1,17 @@
 import React from 'react';
-import BoxList from '../components/layout/Box/BoxList';
-import MediaVideo from '../components/features/MediaVideo/MediaVideo.jsx';
-import MediaPicture from '../components/features/MediaPicture/MediaPicture.jsx';
 
-import regionalImage from '../assets/images/regional copy.png';
-import CiutatVella from '../assets/images/nacional.jpg';
-import zeldaVideo from '../assets/video/zelda.mp4';
-import OIAI from '../assets/video/OIAI.mp4';
-import TextBox from '../components/layout/Box/TextBox.jsx';
+// Componentes
+import BoxList from '../../components/layout/Box/BoxList.jsx';
+import MediaVideo from '../../components/features/MediaVideo/MediaVideo.jsx';
+import MediaPicture from '../../components/features/MediaPicture/MediaPicture.jsx';
+import TextBox from '../../components/layout/Box/TextBox.jsx';
+
+// Imágenes
+import regionalImage from '../../assets/images/regional copy.png';
+import CiutatVella from '../../assets/images/nacional.jpg';
+import zeldaVideo from '../../assets/video/zelda.mp4';
+import OIAI from '../../assets/video/OIAI.mp4';
+
 
 const Services = () => {
     return (
