+++ transportes-fernandez/src/screens/Navigation/Team.jsx
@@ -1,12 +1,14 @@
 import React from 'react';
-import EmployeeCard from '../components/features/Employee/EmployeeCard';
-import narutoImage from '../assets/images/naruto.jpg';
-import letiziaImage from '../assets/images/letizia.png';
-import roshiImage from '../assets/images/roshi.png';
-import marisaImage from '../assets/images/marisa.png';
-import employees from '../data/employees';
-import Background from '../components/features/Background/Background';
-import TextBox from '../components/layout/Box/TextBox';
+import EmployeeCard from '../../components/features/Employee/EmployeeCard';
+import Background from '../../components/features/Background/Background';
+import TextBox from '../../components/layout/Box/TextBox';
+
+import narutoImage from '../../assets/images/naruto.jpg';
+import letiziaImage from '../../assets/images/letizia.png';
+import roshiImage from '../../assets/images/roshi.png';
+import marisaImage from '../../assets/images/marisa.png';
+import employees from '../../data/employees';
+
 
 
 const Team = () => {
